+++ src/application/confirmation-popin/example/index.html
@@ -30,7 +30,7 @@
   <div id='anchor'></div>
   <script type="text/jsx">
     var ConfirmationPopin = FocusComponents.application.confirmationPopin.component;
-    var Button = FocusComponents.common.button.action.component;
+    var {Button} = FocusComponents.components;
 
     var Wrapper = React.createClass({
       getInitialState() {
